Add an `UnicodeCoercingText` TypeDecorator for the `Text` type that
always returns unicode for values fetched from the DB. This allows
us to have a guarantee to always receive unicode for `Text` types,
even when using Oracle as the backend.

The effect of this TypeDecorator should be the same as the cxOracle
dialect option `coerce_to_unicode`, which unfortunately only affects
`String` columns.